Update PersonLinkSchema and EventPersonSchema
Replace person link serialization methods with the respective schema.
