+++ indico/modules/events/schemas.py
@@ -5,13 +5,11 @@
 # modify it under the terms of the MIT License; see the
 # LICENSE file for more details.
 
-from marshmallow import EXCLUDE, fields
+from marshmallow import fields
 
 from indico.core.marshmallow import mm
 from indico.modules.events.models.events import Event
 from indico.modules.events.models.principals import EventPrincipal
-from indico.modules.users.models.users import UserTitle
-from indico.util.i18n import orig_string
 from indico.util.marshmallow import PrincipalPermissionList
 
 
@@ -25,31 +23,6 @@ class EventDetailsSchema(mm.SQLAlchemyAutoSchema):
         fields = ('id', 'category_chain', 'title', 'start_dt', 'end_dt')
 
     category_chain = fields.List(fields.String(), attribute='category.chain_titles')
-
-
-# TODO: move to /persons
-class PersonLinkSchema(mm.Schema):
-    class Meta:
-        unknown = EXCLUDE
-
-    _type = fields.Constant('PersonLink')
-    title = fields.Method('get_title', deserialize='load_title')
-    first_name = fields.String(missing='')
-    last_name = fields.String(required=True)
-    affiliation = fields.String(missing='')
-    phone = fields.String(missing='')
-    address = fields.String(missing='')
-    email = fields.String(required=True)
-    display_order = fields.Int()
-    user_id = fields.Function(lambda o: o.person.user_id)
-    name = fields.String(data_key='display_full_name')
-    avatar_url = fields.Function(lambda o: o.person.user.avatar_url)
-
-    def get_title(self, obj):
-        return obj.title
-
-    def load_title(self, title):
-        return next((x.value for x in UserTitle if title == orig_string(x.title)), UserTitle.none)
 
 
 event_permissions_schema = EventPermissionsSchema()
