+++ indico/modules/events/sessions/fields.py
@@ -7,7 +7,6 @@
 
 from indico.modules.events.fields import PersonLinkListFieldBase
 from indico.modules.events.sessions.models.persons import SessionBlockPersonLink
-from indico.modules.events.util import serialize_person_link
 from indico.web.forms.widgets import JinjaWidget
 
 
@@ -16,8 +15,9 @@ class SessionBlockPersonLinkListField(PersonLinkListFieldBase):
     linked_object_attr = 'session_block'
     widget = JinjaWidget('events/sessions/forms/session_person_link_widget.html')
 
-    def _serialize_person_link(self, principal, extra_data=None):
-        return (extra_data or {}) | serialize_person_link(principal)
+    def _serialize_person_link(self, principal):
+        from indico.modules.events.persons.schemas import PersonLinkSchema
+        return PersonLinkSchema().dump(principal)
 
     def _convert_data(self, data):
         return list({self._get_person_link(x) for x in data})
